+++ ivp/src/lib_marineview/Shape_Kayak.h
@@ -23,7 +23,7 @@
 #ifndef SHAPE_KAYAK_HEADER
 #define SHAPE_KAYAK_HEADER
 
-float g_kayakBody[]=
+double g_kayakBody[]=
 {
     0.0,      0.0,     
     5.25,     6.0,
@@ -83,9 +83,9 @@ float g_kayakBody[]=
    0.0,       0.0
 };
 unsigned int g_kayakBodySize = 58;
-float        g_kayakScale = 1.0;
+double        g_kayakScale = 1.0;
 
-float g_kayakMidOpen[]=
+double g_kayakMidOpen[]=
 {
     0.0,     37.5,     
     9.75,    40.0,
